feat: update task component

+++ resources/views/livewire/tasks/task.blade.php
@@ -11,13 +11,21 @@ new class extends Component
 {
     public Task $task;
 
+    protected $listeners = [
+        'task-updated' => '$refresh',
+    ];
+
     public function completeTask():void
     {
         $this->task->complete();
+
+        $this->dispatch('task-completed');
     }
 
     public function deleteTask():void
     {
+        $this->authorize('delete', Auth::user());
+
         $this->task->delete();
 
         $this->dispatch('task-deleted');
@@ -31,7 +39,7 @@ new class extends Component
         <x-priority-lable :priority="$task->priority"></x-priority-lable>
     </div>
     @if ($task->status)
-        <div class="flex flex-row justify-end items-center mt-5">
+        <div class="flex flex-row justify-end items-center mt-3">
             <p class="text-s">Status :</p>
             <div class="ml-2 px-3 py-1 bg-gray-500 rounded-full text-white text-xs">
                 {{ $task->status->name }}
@@ -45,6 +53,9 @@ new class extends Component
         <div>
             Due Date: {{ $task->due_at ? $task->due_at->format('d-m-Y') : 'Not set' }}
         </div>
+        @if($task->isCompleted())
+            Completed at: {{ $task->completed_at->format('d-m-Y') }}
+        @endif
         <div>
             @if(! $task->isCompleted())
                 <x-primary-button wire:click.prevent="completeTask">
@@ -52,7 +63,9 @@ new class extends Component
                         <path stroke-linecap="round" stroke-linejoin="round" d="m4.5 12.75 6 6 9-13.5" />
                     </svg>
                 </x-primary-button>
-                <x-secondary-button>
+                <x-secondary-button
+                    x-data=""
+                    x-on:click.prevent="$dispatch('open-modal', 'update-task')">
                     <svg class="w-4" xmlns="http://www.w3.org/2000/svg" fill="none" viewBox="0 0 24 24" stroke-width="1.5" stroke="currentColor">
                         <path stroke-linecap="round" stroke-linejoin="round" d="m16.862 4.487 1.687-1.688a1.875 1.875 0 1 1 2.652 2.652L10.582 16.07a4.5 4.5 0 0 1-1.897 1.13L6 18l.8-2.685a4.5 4.5 0 0 1 1.13-1.897l8.932-8.931Zm0 0L19.5 7.125M18 14v4.75A2.25 2.25 0 0 1 15.75 21H5.25A2.25 2.25 0 0 1 3 18.75V8.25A2.25 2.25 0 0 1 5.25 6H10" />
                     </svg>
@@ -65,4 +78,5 @@ new class extends Component
             </x-danger-button>
         </div>
     </div>
+    <livewire:tasks.update :task="$task" />
 </li>
